refactor(pi): modularize user extension features
