+++ files/pi/agent/extensions/user/features/quick-actions.ts
@@ -4,9 +4,8 @@ import type {
 } from "@earendil-works/pi-coding-agent";
 import type { Feature } from "../feature";
 import { showModelSelector } from "../lib/model";
-import { showQuickActions } from "../lib/quick-actions";
+import { runQuickAction, showQuickActions } from "../lib/quick-actions";
 import { showEffortSelector } from "../lib/thinking";
-import { showFocusQuickAction } from "./focus";
 
 const openQuickActions =
 	(pi: ExtensionAPI) =>
@@ -14,7 +13,7 @@ const openQuickActions =
 		const action = await showQuickActions(ctx);
 		switch (action) {
 			case "focus":
-				await showFocusQuickAction(ctx);
+				await runQuickAction("focus", ctx);
 				break;
 			case "effort":
 				await showEffortSelector(pi, ctx);
@@ -32,4 +31,8 @@ function register(pi: ExtensionAPI): void {
 	});
 }
 
-export default { name: "quick-actions", register } satisfies Feature;
+export function createQuickActionsFeature(): Feature {
+	return { name: "quick-actions", dependsOn: ["focus"], register };
+}
+
+export default createQuickActionsFeature();
